refactor(socials): extract ingest debug control plane
